Module and file tree reorganization
  * removed toplevel rendering and audio modules, for reduction
    of noise
  * added ratr0.h which simplifies usage from games
  * moved example-01 images under the assets and orig_assets
    directories
  * use scene node factory to create sprite
  * Centipede:
    * add module and build
    * added sprite load for head
  * refactoring: merged Amiga code
    * Amiga specific code is merged into main engine. It does not
      make much sense to make very generic engine code
  * removed some of the _amiga_ and Amiga names

+++ src/resources.c
@@ -125,7 +125,7 @@ UINT32 ratr0_read_spritesheet(const char *filename, struct Ratr0SpriteSheet *she
         // 2. read sprite colors
         elems_read = fread(sheet->colors, sizeof(UINT16), palette_size, fp);
         for (int i = 0; i < sheet->header.num_sprites; i++) {
-            printf("offset[%d]: %d\n", i, sheet->sprite_offsets[i]);
+            printf("offset[%d]: %d\n", i, (int) sheet->sprite_offsets[i]);
         }
 
         // 3. read image data
